tests/requirements-test.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-flask>=1.2.0 